Overhaul backend structure
