+++ backend/app/config.py
@@ -1,0 +1,1 @@
+# Optional configuration
