+++ backend/app/models.py
@@ -1,0 +1,2 @@
+# Optional for if we decide to use SQLAlchemy 
+# (might be helpful if nobody has major SQL experience)
